Fix the bug where the encryptPrivateKey is repeatedly called when creating an account

+++ lib/core/network/eventCache.dart
@@ -1,4 +1,5 @@
 import 'package:chuchu/core/network/eventDB_isar.dart';
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:isar/isar.dart';
 
 import '../database/db_isar.dart';
@@ -30,7 +31,7 @@ class EventCache {
       cacheIds.add(eventDB.eventId);
     }
 
-    if (expiredEvents.isEmpty) return;
+    if (expiredEvents.isEmpty || kIsWeb) return;
     DBISAR.sharedInstance.isar.write((isar) async {
       int result = await DBISAR.sharedInstance.isar.eventDBISARs.deleteAll(expiredEvents);
       LogUtils.v(() => 'Deleted event caches: $result');
@@ -45,6 +46,7 @@ class EventCache {
   }
 
   Future<void> saveEventToDB(EventDBISAR eventDB) async {
+    if (kIsWeb) return;
     await DBISAR.sharedInstance.saveToDB(eventDB);
   }
